fix product repository implements of mapper struct

+++ src/main/java/com/market/persistence/crud/ProductoCrudRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductoCrudRepository extends CrudRepository<Producto, Long> {
+
+public interface ProductoCrudRepository extends CrudRepository<Producto, Integer> {
 
     List<Producto> findByIdCategoriaOrderByNombreAsc(int idCategoria);
